[Pg] refactor: rename `utils.ts` to `ewkb.ts`
